+++ src/main/java/org/numenta/nupic/Parameters.java
@@ -23,18 +23,23 @@
 package org.numenta.nupic;
 
 import java.io.IOException;
+<<<<<<< HEAD
 import java.util.Arrays;
+=======
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 import java.util.Collections;
-import java.util.EnumMap;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
+import java.util.List;
 import java.util.Random;
 import java.util.Set;
+import java.util.EnumMap;
+import java.util.Arrays;
 
 import org.numenta.nupic.algorithms.SpatialPooler;
 import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.model.Cell;
+import org.numenta.nupic.model.Segment;
 import org.numenta.nupic.model.Column;
 import org.numenta.nupic.model.ComputeCycle;
 import org.numenta.nupic.model.DistalDendrite;
@@ -417,8 +422,10 @@ public class Parameters implements Persistable {
         
         // Network Layer indicator for auto classifier generation
         AUTO_CLASSIFY("hasClassifiers", Boolean.class),
-        
-        
+
+        /** Maps encoder input field name to type of classifier to be used for them */
+        INFERRED_FIELDS("inferredFields", Map.class), // Map<String, Classifier.class>
+
         // How many bits to use if encoding the respective date fields.
         // e.g. Tuple(bits to use:int, radius:double)
         DATEFIELD_SEASON("season", Tuple.class), 
@@ -795,6 +802,7 @@ public class Parameters implements Persistable {
      * The maximum number of synapses added to a segment during learning.
      *
      * @param maxSynapsesPerSegment
+<<<<<<< HEAD
      */
     public void setMaxSynapsesPerSegment(int maxSynapsesPerSegment) {
         paramMap.put(KEY.MAX_SYNAPSES_PER_SEGMENT, maxSynapsesPerSegment);
@@ -805,6 +813,18 @@ public class Parameters implements Persistable {
      *
      * @param maxSegmentsPerCell
      */
+=======
+     */
+    public void setMaxSynapsesPerSegment(int maxSynapsesPerSegment) {
+        paramMap.put(KEY.MAX_SYNAPSES_PER_SEGMENT, maxSynapsesPerSegment);
+    }
+    
+    /**
+     * The maximum number of {@link Segment}s a {@link Cell} can have.
+     *
+     * @param maxSegmentsPerCell
+     */
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
     public void setMaxSegmentsPerCell(int maxSegmentsPerCell) {
         paramMap.put(KEY.MAX_SEGMENTS_PER_CELL, maxSegmentsPerCell);
     }