+++ src/main/java/org/numenta/nupic/algorithms/SDRClassifier.java
@@ -96,7 +96,11 @@ import gnu.trove.list.array.TIntArrayList;
  * @author David Ray
  * @author Andrew Dillon
  */
+<<<<<<< HEAD
 public class SDRClassifier implements Persistable {
+=======
+public class SDRClassifier implements Persistable, Classifier {
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 	private static final long serialVersionUID = 1L;
 
 	int verbosity = 0;