+++ src/test/java/org/numenta/nupic/network/NetworkTest.java
@@ -42,6 +42,7 @@ import org.numenta.nupic.Parameters;
 import org.numenta.nupic.Parameters.KEY;
 import org.numenta.nupic.algorithms.Anomaly;
 import org.numenta.nupic.algorithms.Anomaly.Mode;
+import org.numenta.nupic.algorithms.CLAClassifier;
 import org.numenta.nupic.algorithms.SpatialPooler;
 import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.datagen.ResourceLocator;
@@ -56,6 +57,7 @@ import org.numenta.nupic.network.sensor.SensorParams;
 import org.numenta.nupic.network.sensor.SensorParams.Keys;
 import org.numenta.nupic.util.FastRandom;
 import org.numenta.nupic.util.MersenneTwister;
+import static org.numenta.nupic.network.NetworkTestHarness.*;
 
 import rx.Observer;
 import rx.Subscriber;
@@ -209,7 +211,12 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
         
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         // Create a Network
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
@@ -270,7 +277,12 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
         
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         // Create a Network
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
@@ -490,7 +502,12 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
         
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         // Create a Network
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
@@ -555,6 +572,10 @@ public class NetworkTest extends ObservableTestBase {
         p.setPotentialRadius(16);
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
@@ -694,6 +715,10 @@ public class NetworkTest extends ObservableTestBase {
         p.set(KEY.MAX_BOOST, 1.0);
         p.set(KEY.DUTY_CYCLE_PERIOD, 7);
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         
         Map<String, Object> params = new HashMap<>();
         params.put(KEY_MODE, Mode.PURE);
@@ -771,6 +796,10 @@ public class NetworkTest extends ObservableTestBase {
         p.set(KEY.MAX_BOOST, 10.0);
         p.set(KEY.DUTY_CYCLE_PERIOD, 7);
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         
         Map<String, Object> params = new HashMap<>();
         params.put(KEY_MODE, Mode.PURE);
@@ -819,6 +848,10 @@ public class NetworkTest extends ObservableTestBase {
         p.set(KEY.MAX_BOOST, 10.0);
         p.set(KEY.DUTY_CYCLE_PERIOD, 7);
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         
         Map<String, Object> params = new HashMap<>();
         params.put(KEY_MODE, Mode.PURE);
@@ -847,6 +880,7 @@ public class NetworkTest extends ObservableTestBase {
         //////////////////////////////////////////////////////
         p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
         n = Network.create("test network", p)
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("1", p)
@@ -869,6 +903,7 @@ public class NetworkTest extends ObservableTestBase {
         try {
             p = NetworkTestHarness.getParameters();
             p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
+            p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
             n = Network.create("test network", p)
                 .add(Network.createRegion("r1")
                     .add(Network.createLayer("1", p)
@@ -965,6 +1000,10 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p = p.union(NetworkTestHarness.getGeospatialTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         HTMSensor<ObservableSensor<String[]>> htmSensor = (HTMSensor<ObservableSensor<String[]>>)sensor;
 
@@ -1050,7 +1089,12 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
         
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("2", p)
@@ -1078,7 +1122,12 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
         
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("2", p)
@@ -1278,7 +1327,12 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p = p.union(NetworkTestHarness.getDayDemoTestEncoderParams());
         p.set(KEY.RANDOM, new FastRandom(42));
+<<<<<<< HEAD
         
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         Sensor<ObservableSensor<String[]>> sensor = Sensor.create(
             ObservableSensor::create, SensorParams.create(Keys::obs, new Object[] {"name", 
                 PublisherSupplier.builder()