+++ src/test/java/org/numenta/nupic/network/PersistenceAPITest.java
@@ -30,6 +30,10 @@ import static org.junit.Assert.fail;
 import static org.numenta.nupic.algorithms.Anomaly.KEY_MODE;
 import static org.numenta.nupic.algorithms.Anomaly.KEY_USE_MOVING_AVG;
 import static org.numenta.nupic.algorithms.Anomaly.KEY_WINDOW_SIZE;
+<<<<<<< HEAD
+=======
+import static org.numenta.nupic.network.NetworkTestHarness.*;
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
 import java.io.File;
 import java.math.BigDecimal;
@@ -48,6 +52,10 @@ import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 import org.junit.AfterClass;
+<<<<<<< HEAD
+=======
+import org.junit.BeforeClass;
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 import org.junit.Test;
 import org.numenta.nupic.FieldMetaType;
 import org.numenta.nupic.Parameters;
@@ -96,6 +104,16 @@ public class PersistenceAPITest extends ObservableTestBase {
     /** Printer to visualize DayOfWeek printouts - SET TO TRUE FOR PRINTOUT */
     private BiFunction<Inference, Integer, Integer> dayOfWeekPrintout = createDayOfWeekInferencePrintout(false);
     
+<<<<<<< HEAD
+=======
+    
+    @BeforeClass
+    public static void beforeClass(){
+    	// Sample data contains datetimes that are invalid in some timezones due to DST.
+    	// If UTC is forced, then test runs should yield the same result regardless of timezone
+    	System.setProperty("user.timezone", "UTC");
+    }
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
     @AfterClass
     public static void cleanUp() {
@@ -109,8 +127,11 @@ public class PersistenceAPITest extends ObservableTestBase {
                         catch(Exception io) { throw new RuntimeException(io); } 
                     }
                 );
+<<<<<<< HEAD
                 
                 Files.delete(serialDir.toPath());
+=======
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
             }
         }catch(Exception e) {
             e.printStackTrace();
@@ -676,6 +697,10 @@ public class PersistenceAPITest extends ObservableTestBase {
     public void testSerializeLayer() {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         Map<String, Map<String, Object>> settings = NetworkTestHarness.setupMap(
             null, // map
             8,    // n
@@ -1684,6 +1709,10 @@ public class PersistenceAPITest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getDayDemoTestEncoderParams());
         p.set(KEY.RANDOM, new FastRandom(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         
         Layer<?> l2 = null;
         Network network = Network.create("test network", p)
@@ -1711,6 +1740,10 @@ public class PersistenceAPITest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p = p.union(NetworkTestHarness.getDayDemoTestEncoderParams());
         p.set(KEY.RANDOM, new FastRandom(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         Sensor<ObservableSensor<String[]>> sensor = Sensor.create(
             ObservableSensor::create, SensorParams.create(Keys::obs, new Object[] {"name", 
@@ -1734,6 +1767,10 @@ public class PersistenceAPITest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         p = p.union(NetworkTestHarness.getNetworkDemoTestEncoderParams());
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
@@ -1759,6 +1796,10 @@ public class PersistenceAPITest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p = p.union(NetworkTestHarness.getHotGymTestEncoderParams());
         p.set(KEY.RANDOM, new FastRandom(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         Sensor<ObservableSensor<String[]>> sensor = Sensor.create(
             ObservableSensor::create, SensorParams.create(Keys::obs, new Object[] {"name", 
@@ -1782,7 +1823,12 @@ public class PersistenceAPITest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p = p.union(NetworkTestHarness.getHotGymTestEncoderParams());
         p.set(KEY.RANDOM, new FastRandom(42));
+<<<<<<< HEAD
         
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
         Network network = Network.create("test network", p).add(Network.createRegion("r1")
             .add(Network.createLayer("1", p)
                 .alterParameter(KEY.AUTO_CLASSIFY, true)
@@ -1797,6 +1843,10 @@ public class PersistenceAPITest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p = p.union(NetworkTestHarness.getHotGymTestEncoderParams());
         p.set(KEY.RANDOM, new FastRandom(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         Object[] n = { "some name", ResourceLocator.path("rec-center-hourly.csv") };
         HTMSensor<File> sensor = (HTMSensor<File>)Sensor.create(
@@ -1824,6 +1874,10 @@ public class PersistenceAPITest extends ObservableTestBase {
 
         Parameters p = NetworkTestHarness.getParameters().copy();
         p.set(KEY.RANDOM, new MersenneTwister(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         Map<String, Map<String, Object>> settings = NetworkTestHarness.setupMap(
             null, // map
@@ -1911,6 +1965,10 @@ public class PersistenceAPITest extends ObservableTestBase {
             ObservableSensor::create, SensorParams.create(Keys::obs, new Object[] {"name", manual}));
 
         Parameters p = getParameters();
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("dayOfWeek", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         Map<String, Map<String, Object>> settings = NetworkTestHarness.setupMap(
             null, // map