+++ src/test/java/org/numenta/nupic/serialize/HTMObjectInputOutputTest.java
@@ -1,5 +1,10 @@
 package org.numenta.nupic.serialize;
 
+<<<<<<< HEAD
+=======
+import static org.numenta.nupic.network.NetworkTestHarness.getInferredFieldsMap;
+
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
@@ -11,6 +16,10 @@ import org.junit.Test;
 import org.numenta.nupic.Parameters;
 import org.numenta.nupic.Parameters.KEY;
 import org.numenta.nupic.algorithms.Anomaly;
+<<<<<<< HEAD
+=======
+import org.numenta.nupic.algorithms.CLAClassifier;
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 import org.numenta.nupic.algorithms.SpatialPooler;
 import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.network.Network;
@@ -58,6 +67,10 @@ public class HTMObjectInputOutputTest {
         Parameters p = NetworkTestHarness.getParameters().copy();
         p = p.union(NetworkTestHarness.getHotGymTestEncoderParams());
         p.set(KEY.RANDOM, new FastRandom(42));
+<<<<<<< HEAD
+=======
+        p.set(KEY.INFERRED_FIELDS, getInferredFieldsMap("consumption", CLAClassifier.class));
+>>>>>>> 8fc6b596461a879fdf3e8936833c9a972d858b57
 
         Sensor<ObservableSensor<String[]>> sensor = Sensor.create(
             ObservableSensor::create, SensorParams.create(Keys::obs, new Object[] {"name",